requirements.txt
Scrapy >= 1.4
kafka-python >= 2.0
